requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
bcrypt
gunicorn
psycopg2-binary
